requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.104.1
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.24.0
